refactor: extract CdpTargetManager
